requirements.txt
streamlit
langchain>=0.1.17
openai
faiss-cpu
python-dotenv
openpyxl
pandas
torch==2.7.1
sentence-transformers==4.1.0
transformers==4.52.4
huggingface-hub==0.32.4
tokenizers==0.21.1
pinecone           # current package name
watchdog
keras==2.11.0
streamlit-option-menu
streamlit-aggrid==0.3.4.post3